+++ async/vault-async/src/main/java/co/com/bancolombia/secretsmanager/connector/VaultSecretManagerConfigurator.java
@@ -1,48 +1,33 @@
 package co.com.bancolombia.secretsmanager.connector;
 
 import co.com.bancolombia.secretsmanager.api.exceptions.SecretException;
-import co.com.bancolombia.secretsmanager.vault.config.VaultKeyStoreProperties;
+import co.com.bancolombia.secretsmanager.vault.K8sTokenReader;
+import co.com.bancolombia.secretsmanager.vault.VaultSecretManagerConfiguratorBase;
 import co.com.bancolombia.secretsmanager.vault.config.VaultSecretsManagerProperties;
-import co.com.bancolombia.secretsmanager.vault.config.VaultTrustStoreProperties;
-import co.com.bancolombia.secretsmanager.vault.ssl.SslConfig;
 import lombok.Builder;
 
 import java.net.http.HttpClient;
-import java.time.Duration;
-import java.util.logging.Logger;
 
 /**
- * This class is in charge of configuring the VaultSecretsManagerConnector
+ * This class is in charge of configuring the Async VaultSecretsManagerConnector
  */
-@Builder(setterPrefix = "with", toBuilder = true)
-public class VaultSecretManagerConfigurator {
+public class VaultSecretManagerConfigurator extends VaultSecretManagerConfiguratorBase {
 
-    private static final Logger logger = Logger.getLogger("config.VaultSecretManagerConfigurator");
-    private final VaultSecretsManagerProperties properties;
     private final K8sTokenReader k8sTokenReader;
 
-    /**
-     * This method is in charge of configuring the HttpClient
-     * @return HttpClient configured.
-     * @throws SecretException
-     */
-    public HttpClient getHttpClient() throws SecretException {
-        HttpClient.Builder clientBuilder = HttpClient.newBuilder()
-                .followRedirects(HttpClient.Redirect.NORMAL)
-                .connectTimeout(Duration.ofSeconds(properties.getHttpProperties().getConnectionTimeout()));
-        if (properties.isSsl() || properties.getTrustStoreProperties() != null || properties.getKeyStoreProperties() != null) {
-            clientBuilder = clientBuilder.sslContext(buildSslConfig().getSslContext());
-        }
-        return clientBuilder.build();
+    @Builder(setterPrefix = "with", toBuilder = true)
+    public VaultSecretManagerConfigurator(VaultSecretsManagerProperties properties, K8sTokenReader k8sTokenReader) {
+        super(properties);
+        this.k8sTokenReader = k8sTokenReader;
     }
 
     /**
      * This method is in charge of configuring the VaultAuthenticator
      * @return the VaultAuthenticator configured.
-     * @throws SecretException
+     * @throws SecretException when an error occurs while reading the kubernetes service account token
      */
     public VaultAuthenticator getVaultAuthenticator() throws SecretException {
-        return new VaultAuthenticator(getHttpClient(), properties,
+        return new VaultAuthenticator(getHttpClient(), getProperties(),
                 k8sTokenReader != null? k8sTokenReader : new K8sTokenReader());
     }
 
@@ -55,43 +40,7 @@ public class VaultSecretManagerConfigurator {
         HttpClient httpClient = getHttpClient();
         return new VaultSecretsManagerConnectorAsync(httpClient,
                 getVaultAuthenticator(),
-                properties);
+                getProperties());
     }
 
-    private SslConfig buildSslConfig() throws SecretException {
-        SslConfig sslConfig = new SslConfig();
-        if (properties.getTrustStoreProperties() != null) {
-            setTrustConfiguration(sslConfig, properties.getTrustStoreProperties());
-        }
-        if (properties.getKeyStoreProperties() != null) {
-            setKeystoreConfiguration(sslConfig, properties.getKeyStoreProperties());
-        }
-        return sslConfig.build();
-    }
-
-    private SslConfig setTrustConfiguration(SslConfig sslConfig,
-                                            VaultTrustStoreProperties trustStoreProperties) throws SecretException {
-        if (trustStoreProperties.getTrustStoreJksFile() != null) {
-            sslConfig.trustStoreFile(trustStoreProperties.getTrustStoreJksFile());
-        }  else if (trustStoreProperties.getPemFile() != null) {
-            sslConfig.pemFile(trustStoreProperties.getPemFile());
-        } else {
-            throw new SecretException("VaultTrustStoreProperties was set, but no trust store file or pem resource provided");
-        }
-        return sslConfig;
-    }
-
-    private SslConfig setKeystoreConfiguration(SslConfig sslConfig,
-                                               VaultKeyStoreProperties keyStoreProperties) throws SecretException {
-        if (keyStoreProperties.getKeyStoreFile() != null) {
-            sslConfig.keyStoreFile(keyStoreProperties.getKeyStoreFile(),
-                    keyStoreProperties.getKeyStorePassword());
-        } else if (keyStoreProperties.getClientPem() != null && keyStoreProperties.getClientKeyPem() != null) {
-            sslConfig.clientPemFile(keyStoreProperties.getClientPem());
-            sslConfig.clientKeyPemFile(keyStoreProperties.getClientKeyPem());
-        } else {
-            throw new SecretException("VaultKeyStoreProperties was set, but no key store file or pem resources provided");
-        }
-        return sslConfig;
-    }
 }
